invconv: Move files up one folder from 'modules'
Also changed vscode settings.json so parser won't complain about import
files not existing on some machines.

As well, all instances of cell_loc were replaced with cell_pos because
the latter is a more intuitive name.
