+++ .vscode/settings.json
@@ -3,5 +3,8 @@
     "python.linting.prospectorEnabled": true,
     "python.linting.pylintEnabled": false,
     "python.linting.pylintUseMinimalCheckers": false,
-    "editor.formatOnSave": true
+    "editor.formatOnSave": true,
+    "python.analysis.extraPaths": [
+        "./invconv"
+    ]
 }
