--- invconv/modules/__init__.py
@@ -1,3 +1,0 @@
-# Copyright 2021 Richard Johnston <techpowerawaits@outlook.com>
-# SPDX-license-identifier: 0BSD
-pass
